feat: add Webtoons' API abstraction
Current structure is a little confusing with a `bin/` folder to hold
source code, as well as the nesting underneath it, with lots of
unnecessary standalone files, which as an approach should really be used
to hold appropriate abstractions. The refactor will start to use the new
`src/` folder as the new entry point.

With a flatter structure, the hope is to improve localized awareness
when trying to parse through code, making it see easier to reason about
what is happening, and hopefully making contributions from the
community, who may not have great programming knowledge, but if it was
easier to reason about, a greater possibility.

This first part begins the abstraction over webtoons API, wrapping it in
a class, with the goal keeping this domain separate from injecting into
the DOM. With having the full API for posts returned, it should also
allow others to use a simple object to add new bits of information
without needing domain knowledge of the API.