feat(webtoon): add `getTodaysOrNewestPosts` method
This is a specialized method intended to be run first for any webtoon.
This will get all of the days comments, and falling back from that, the
newest comment for each episode.

The use case is to help deal with issues of pathological amounts of
comments on some originals. localStorage only has 10MB of storage space
alloted for it, so it would be impossible to store 2 million comments as
seen on unOrdinary. Given the main use case for Comment Wizard,
freshness, we can optimize getting only the newest comments for each
episode. With this method as the starter, the slate can start fresh and
can stay fast.