feat(post): implement `delete`
When exploring, there was actually a slight difference in what deleting
your own post looked liked vs a creator being able to delete the post of
someone else.

The URL endpoint to delete is the same, but the requirement of the API
token is needed to delete someone else's. Given that the prominent use
will be to delete other people's posts, this implementation includes the
API token. But no tests have been done to see if this would also work
for a creator deleting their post/reply.